Don't migrate more companies after some went live
It is not supported to migrate some companies, then go live on them, then migrate some more. This has led to incidents and data corruption. Adding this statement to ensure that the partners are aware of this. 
Also removed the step for setting up the periodic sync - we no longer have that step.