new paragraph about telemetry cost (#3831)
* new paragraph about telemetry cost

* Update telemetry-overview.md

* Queries in Monitoring and Analyzing Telemetry

Not sure if this is for John or Kennie....
I fixed some typos and made other changes for easier reading and simplicity of expression. I also put a few queries in, and I would love to know the answers to them as that will help me learn! Thanks.

* Answers to Sarah's questions

* Edit review of Telemetry paragraph

Hi John -- I accepted your comments so the article is now clear of corrections. Note that I deleted everything but the first sentence under "Available telemetry" -- just wanted to double-check that with you (that was where we had that hidden content). Thanks!

Co-authored-by: jswymer <jswymer@microsoft.com>
Co-authored-by: Sarah Hersack <107584989+sh1801@users.noreply.github.com>