Changing date after failure can be blocked for 24h
It is very important to note that after a failed update, if users need to reschedule the update, they cannot to this immediately or after few hours. Unblocking scheduled updates happens typically once per day and early in the morning. 
CSS is receiving quite frequently requests like yesterday I couldn't reschedule, but today I can. 
This is the reason why it is important to add this in the documentation.