Remark about OnFindRecord and OnNextRecord
If OnFindRecord and/or OnNextRecord are just defined in the List/ListPart (even if they do not have any code in it), platform will not apply any set load fields to load. 
See also this request to refactor Item list page to improve performance.
https://dynamicssmb2.visualstudio.com/Dynamics%20SMB/_workitems/edit/433399

This remark is to generate awareness about this hidden platform behavior that have impact on performance when designing List/ListParts.