Revisited UseFilters note for better clarity.
The client types mentioned are possibly inccorect. Current refers to Session.CurrentClientType() and would not work within an API page, specifying ClientType::Current. This is the main reason to refactor the sentence. 
Removed also the last part. The text is right regarding the parameter itself, but a false should not be counted as specified. Even false for e.g. GetFilters returns an empty string when using Api.