refer to LC0105 in extensions disappearing FAQ question
We will start emitting a new dimension on partner telemetry event LC0105, which describes the reason for the environment upgrade (whether it's planned or internal maintenance). This change briefly mentions this event (and backref on how to set up partner telemetry) in hopes that partners may be able to confirm this behavior before opening an IcM ticket.