Add more rounding examples to avoid confusion
The examples for `Round()` jump from `Round(..., 100)` directly to `Round(..., 0.1)`. The resulting rounded number (1200) can easily be mistaken for 1234. Hence, I've seen more than one developer trying to use `100` as a parameter to round to integers.

Of course, proper reading could prevent this, but just adding the examples for `1` and `10` (both positive and negative) is just as easy to prevent this confusion.